Move implementation of Occurs to exp.rs
