+++ lang/elaborator/src/typechecker/typecheck.rs
@@ -11,11 +11,11 @@ use miette_util::ToMiette;
 use syntax::common::*;
 use syntax::ctx::values::Binder;
 use syntax::ctx::{BindContext, BindElem, LevelCtx};
-use syntax::generic::{Hole, Named, TypeUniv, Variable};
+use syntax::generic::{Hole, Named, Occurs, TypeUniv, Variable};
 use syntax::tst::forget::ForgetTST;
 use syntax::tst::{self, HasTypeInfo};
 use syntax::ust::util::Instantiate;
-use syntax::ust::{self, Occurs};
+use syntax::ust::{self};
 use tracer::trace;
 
 use super::ctx::*;
