+++ lang/syntax/src/trees/generic/mod.rs
@@ -2,7 +2,9 @@ mod decls;
 mod exp;
 mod lookup;
 pub mod lookup_table;
+mod traits;
 
 pub use decls::*;
 pub use exp::*;
 pub use lookup::*;
+pub use traits::*;
