+++ lang/syntax/src/trees/ust/mod.rs
@@ -1,7 +1,5 @@
 mod def;
-mod occurs;
 mod subst;
 pub mod util;
 
 pub use def::*;
-pub use occurs::*;
