--- lang/syntax/src/trees/ust/occurs.rs
@@ -1,127 +1,0 @@
-use std::rc::Rc;
-
-use crate::common::*;
-use crate::ctx::*;
-use crate::generic::Hole;
-use crate::generic::TypeUniv;
-use crate::generic::Variable;
-
-use super::def::*;
-
-pub fn occurs_in(ctx: &mut LevelCtx, the_idx: Idx, in_exp: &Rc<Exp>) -> bool {
-    let lvl = ctx.idx_to_lvl(the_idx);
-    in_exp.occurs(ctx, lvl)
-}
-
-pub trait Occurs {
-    fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool;
-}
-
-impl Occurs for Exp {
-    fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
-        match self {
-            Exp::Variable(e) => e.occurs(ctx, lvl),
-            Exp::TypCtor(e) => e.occurs(ctx, lvl),
-            Exp::Call(e) => e.occurs(ctx, lvl),
-            Exp::DotCall(e) => e.occurs(ctx, lvl),
-            Exp::Anno(e) => e.occurs(ctx, lvl),
-            Exp::TypeUniv(e) => e.occurs(ctx, lvl),
-            Exp::LocalMatch(e) => e.occurs(ctx, lvl),
-            Exp::LocalComatch(e) => e.occurs(ctx, lvl),
-            Exp::Hole(e) => e.occurs(ctx, lvl),
-        }
-    }
-}
-
-impl Occurs for Variable {
-    fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
-        let Variable { idx, .. } = self;
-        ctx.idx_to_lvl(*idx) == lvl
-    }
-}
-
-impl Occurs for TypCtor {
-    fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
-        let TypCtor { args, .. } = self;
-        args.args.iter().any(|arg| arg.occurs(ctx, lvl))
-    }
-}
-
-impl Occurs for Call {
-    fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
-        let Call { args, .. } = self;
-        args.args.iter().any(|arg| arg.occurs(ctx, lvl))
-    }
-}
-
-impl Occurs for DotCall {
-    fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
-        let DotCall { exp, args, .. } = self;
-        exp.occurs(ctx, lvl) || args.args.iter().any(|arg| arg.occurs(ctx, lvl))
-    }
-}
-
-impl Occurs for Anno {
-    fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
-        let Anno { exp, typ, .. } = self;
-        exp.occurs(ctx, lvl) || typ.occurs(ctx, lvl)
-    }
-}
-
-impl Occurs for TypeUniv {
-    fn occurs(&self, _ctx: &mut LevelCtx, _lvl: Lvl) -> bool {
-        false
-    }
-}
-
-impl Occurs for LocalMatch {
-    fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
-        let LocalMatch { on_exp, body, .. } = self;
-        on_exp.occurs(ctx, lvl) || body.occurs(ctx, lvl)
-    }
-}
-
-impl Occurs for LocalComatch {
-    fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
-        let LocalComatch { body, .. } = self;
-        body.occurs(ctx, lvl)
-    }
-}
-
-impl Occurs for Hole {
-    fn occurs(&self, _ctx: &mut LevelCtx, _lvl: Lvl) -> bool {
-        false
-    }
-}
-
-impl Occurs for Match {
-    fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
-        let Match { cases, .. } = self;
-        cases.occurs(ctx, lvl)
-    }
-}
-
-impl Occurs for Case {
-    fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
-        let Case { params, body, .. } = self;
-        ctx.bind_iter(params.params.iter().map(|_| ()), |ctx| body.occurs(ctx, lvl))
-    }
-}
-
-impl<T: Occurs> Occurs for Rc<T> {
-    fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
-        T::occurs(self, ctx, lvl)
-    }
-}
-
-impl<T: Occurs> Occurs for Vec<T> {
-    fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
-        self.iter().any(|x| x.occurs(ctx, lvl))
-    }
-}
-
-impl<T: Occurs> Occurs for Option<T> {
-    fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
-        self.as_ref().map(|inner| inner.occurs(ctx, lvl)).unwrap_or_default()
-    }
-}
